And merge branch 'develop' into thirdparty-submodules

Conflicts:
	utils/installers/rti/linux/setup_linux_fastrpc.sh
	utils/pcPerformance/latency/eprosima_struct/fastrpc/Makefile

+++ utils/installers/rti/linux/setup_linux_fastrpc.sh
@@ -48,7 +48,7 @@ function installer
 
     # Copy example.
     mkdir -p tmp/$project/examples/C++
-    cp -r ../../../../examples/C++/FastRPC tmp/$project/examples/C++
+    cp -r ../../../../examples/C++/FastRPC/* tmp/$project/examples/C++
     errorstatus=$?
     if [ $errorstatus != 0 ]; then return; fi
 
@@ -164,8 +164,10 @@ function installer
 function rpminstaller
 {
     rm tmp/$project/classes/antlr-2.7.7.jar
-    rm tmp/$project/classes/stringtemplate-3.2.1.jar
-    rm tmp/$project/classes/rpcddsgen.jar
+    rm tmp/$project/classes/fastrpcgen.jar
+    if [[ "${distroversion}" == "Fedora"* ]]; then 
+        rm tmp/$project/classes/stringtemplate-3.2.1.jar
+    fi
 
 	# Change the script form local to general script.
 	cp ../../../../scripts/fastrpcgen_rpm.sh tmp/$project/scripts/fastrpcgen.sh
@@ -177,7 +179,7 @@ function rpminstaller
 	mkdir -p tmp/$project/fastrpcgen
 
 	# Copy the build.xml
-	if [ ${distroversion} == CentOS6.4 ]; then
+	if [[ "${distroversion}" == "CentOS6"* ]]; then
 		cp build_rpm_fastrpc_centos.xml tmp/$project/fastrpcgen/build.xml
 	else
 		cp build_rpm_fastrpc.xml tmp/$project/fastrpcgen/build.xml
@@ -215,8 +217,8 @@ function rpminstaller
 	if [ $errorstatus != 0 ]; then return; fi
 
 	# Copy SPEC file
-	if [ ${distroversion} == CentOS6.4 ]; then
-		sed "s/VERSION/${version}/g" FastBuffers_centos.spec > ~/rpmbuild/SPECS/FastBuffers.spec
+	if [[ "${distroversion}" == "CentOS6"* ]]; then
+		sed "s/VERSION/${version}/g" fastrpc_centos.spec > ~/rpmbuild/SPECS/fastrpc.spec
 	else
 		sed "s/VERSION/${version}/g" fastrpc.spec > ~/rpmbuild/SPECS/fastrpc.spec
 	fi
@@ -307,8 +309,11 @@ mkdir tmp/$project
 
 installer
 
-# TODO Detect if the distro suport RPM
-[ $errorstatus == 0 ] && { rpminstaller; }
+if [ $errorstatus == 0 ]; then
+    if [[ "${distroversion}" == "CentOS6"* ]] || [[ "${distroversion}" == "Fedora"* ]]; then
+        rpminstaller
+    fi
+fi
 
 # Remove temporaly directory
 rm -rf tmp
